Move evolve_fitness_trait_pop helpers to own files
