+++ hstrat/test_drive/generate_template_phylogeny/_evolve_fitness_trait_population_/__init__.py
@@ -1,0 +1,7 @@
+from ._apply_island_swaps import _apply_island_swaps
+from ._apply_mutation import _apply_mutation
+from ._apply_niche_swaps import _apply_niche_swaps
+from ._do_selection import _do_selection
+from ._get_island_id import _get_island_id
+from ._get_niche_id import _get_niche_id
+from ._setup_population import _setup_population
