+++ hstrat/test_drive/generate_template_phylogeny/_evolve_fitness_trait_population_/_apply_mutation.py
@@ -1,0 +1,7 @@
+import numpy as np
+
+
+def _apply_mutation(
+    pop_arr: np.array,
+) -> None:
+    pop_arr += np.random.standard_normal(size=len(pop_arr))
