+++ hstrat/test_drive/generate_template_phylogeny/_evolve_fitness_trait_population_/_get_island_id.py
@@ -1,0 +1,9 @@
+import numba as nb
+
+
+@nb.jit(nopython=True)
+def _get_island_id(
+    population_idx: int,
+    island_size: int,
+) -> int:
+    return population_idx // island_size
