+++ hstrat/test_drive/generate_template_phylogeny/_evolve_fitness_trait_population_/_get_niche_id.py
@@ -1,0 +1,8 @@
+import numba as nb
+
+
+@nb.jit(nopython=True)
+def _get_niche_id(
+    population_idx: int, island_niche_size: int, num_niches: int
+) -> int:
+    return (population_idx // island_niche_size) % num_niches
